Fix bug in set_double_attr().
git-svn-id: http://svn.research.cc.gatech.edu/kaos/atl/trunk@3352 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ attr.c
@@ -398,8 +398,8 @@ double dvalue;
     attr_value_type t = Attr_Float4;
     attr_union tmp;
     tmp.u.d = dvalue;
-    if (sizeof(float) == 8) t = Attr_Float8;
-    if (sizeof(float) == 16) t = Attr_Float16;
+    if (sizeof(double) == 8) t = Attr_Float8;
+    if (sizeof(double) == 16) t = Attr_Float16;
     return set_pattr(list, attr_id, t, tmp);
 }
